+++ PEngine.Core.Shared/Models/ResumeObjectiveModel..cs
@@ -6,7 +6,7 @@ namespace PEngine.Core.Shared.Models
   public class ResumeObjectiveModel : IGuidModel, ITimestampModel
   {
     public Guid Guid { get; set; }
-    public Int32? LegacyID { get; set; }
+    public long? LegacyID { get; set; }
     public string Data { get; set; } = string.Empty;
     public DateTime? CreatedUTC { get; set; }
     public DateTime? ModifiedUTC { get; set; }
