+++ PEngine.Core.Shared/Models/ResumeWorkHistoryModel.cs
@@ -6,7 +6,7 @@ namespace PEngine.Core.Shared.Models
   public class ResumeWorkHistoryModel : IGuidModel, ITimestampModel
   {
     public Guid Guid { get; set; }
-    public Int32? LegacyID { get; set; }
+    public long? LegacyID { get; set; }
     public string Employer { get; set; } = string.Empty;
     public string EmployerURL { get; set; } = string.Empty;
     public string JobTitle { get; set; } = string.Empty;
